Updating formatting and fixing compiler warnings

+++ vtkm/cont/testing/UnitTestDataSetBuilderRectilinear.cxx
@@ -162,12 +162,12 @@ RectilinearTests()
     std::cout << "1D cases" << std::endl;
     numPoints = dimensions[0];
     numCells = dimensions[0]-1;
-    std::vector<T> varP1D(numPoints);
+    std::vector<T> varP1D(static_cast<unsigned long>(numPoints));
     for (vtkm::Id i = 0; i < numPoints; i++)
     {
       varP1D[i] = static_cast<T>(i*1.1f);
     }
-    std::vector<T> varC1D(numCells);
+    std::vector<T> varC1D(static_cast<unsigned long>(numCells));
     for (vtkm::Id i = 0; i < numCells; i++)
     {
       varC1D[i] = static_cast<T>(i*1.1f);
@@ -182,12 +182,12 @@ RectilinearTests()
     std::cout << "2D cases" << std::endl;
     numPoints = dimensions[0]*dimensions[1];
     numCells = (dimensions[0]-1)*(dimensions[1]-1);
-    std::vector<T> varP2D(numPoints);
+    std::vector<T> varP2D(static_cast<unsigned long>(numPoints));
     for (vtkm::Id i = 0; i < numPoints; i++)
     {
       varP2D[i] = static_cast<T>(i*1.1f);
     }
-    std::vector<T> varC2D(numCells);
+    std::vector<T> varC2D(static_cast<unsigned long>(numCells));
     for (vtkm::Id i = 0; i < numCells; i++)
     {
       varC2D[i] = static_cast<T>(i*1.1f);
@@ -218,12 +218,12 @@ RectilinearTests()
     std::cout << "3D cases" << std::endl;
     numPoints *= dimensions[2];
     numCells *= dimensions[2]-1;
-    std::vector<T> varP3D(numPoints);
+    std::vector<T> varP3D(static_cast<unsigned long>(numPoints));
     for (vtkm::Id i = 0; i < numPoints; i++)
     {
       varP3D[i] = static_cast<T>(i*1.1f);;
     }
-    std::vector<T> varC3D(numCells);
+    std::vector<T> varC3D(static_cast<unsigned long>(numCells));
     for (vtkm::Id i = 0; i < numCells; i++)
     {
       varC3D[i] = static_cast<T>(i*1.1f);
